Example data for RedoxRRI

A small simulated dataset bundled with RedoxRRI to demonstrate the full workflow of computing the Holobiont Redox Resilience Index (RRI) and visualizing plant–soil–microbe contributions.

Format

A named list with components:

ROS_flux

data.frame; samples (rows) by plant physiological traits (columns).

Eh_stability

data.frame; samples (rows) by soil redox variables (columns).

micro_data

data.frame; samples (rows) by microbial features (columns).

graph

An igraph object representing a microbial association network.

id

data.frame; sample metadata (plot, depth, time).

Source

Simulated data generated for examples and unit tests.

Details

This dataset is synthetic and intended solely for:

Function examples

Unit tests

Vignettes and tutorials '

Values are generated to mimic realistic structure:

Multivariate plant physiological traits (e.g. Fv/Fm, NPQ)

Soil redox chemistry gradients (Eh, Fe/Mn ratios)

High-dimensional, sparse microbial abundance data

A system-level microbial association network

The data do not represent any real biological system.

Examples

data("redoxrri_example")

str(redoxrri_example)
#> List of 5
#>  $ ROS_flux    :'data.frame':	30 obs. of  4 variables:
#>   ..$ FvFm   : num [1:30] 0.849 0.879 0.99 0.889 0.828 ...
#>   ..$ PhiPSII: num [1:30] 0.46 0.466 0.533 0.425 0.376 ...
#>   ..$ NPQ    : num [1:30] 1.21 1.013 0.779 0.9 1.346 ...
#>   ..$ SPAD   : num [1:30] 37.3 35.9 39.9 37.1 41.3 ...
#>  $ Eh_stability:'data.frame':	30 obs. of  4 variables:
#>   ..$ Eh     : num [1:30] 0.78 0.935 1.125 0.716 0.61 ...
#>   ..$ Fe2_Fe3: num [1:30] 0.623 0.874 0.284 0.557 0.861 ...
#>   ..$ Mn2_Mn4: num [1:30] 0.53 0.322 0.124 0.377 0.55 ...
#>   ..$ NH4_NO3: num [1:30] 1.009 0.949 0.899 1.136 1.356 ...
#>  $ micro_data  :'data.frame':	30 obs. of  40 variables:
#>   ..$ ASV1 : num [1:30] 2 3 7 2 1 0 0 0 0 0 ...
#>   ..$ ASV2 : num [1:30] 0 0 5 3 4 1 0 0 2 0 ...
#>   ..$ ASV3 : num [1:30] 2 4 9 3 4 0 1 1 0 0 ...
#>   ..$ ASV4 : num [1:30] 0 0 8 0 1 0 1 1 0 0 ...
#>   ..$ ASV5 : num [1:30] 0 3 8 0 1 1 1 0 1 0 ...
#>   ..$ ASV6 : num [1:30] 1 3 6 2 0 1 0 0 0 1 ...
#>   ..$ ASV7 : num [1:30] 0 0 8 6 0 0 2 0 0 1 ...
#>   ..$ ASV8 : num [1:30] 2 1 4 0 2 1 0 0 0 1 ...
#>   ..$ ASV9 : num [1:30] 0 2 3 1 1 0 0 0 0 0 ...
#>   ..$ ASV10: num [1:30] 0 0 6 0 1 1 0 0 0 0 ...
#>   ..$ ASV11: num [1:30] 2 0 11 1 0 2 0 0 0 0 ...
#>   ..$ ASV12: num [1:30] 3 6 6 1 0 0 0 0 0 0 ...
#>   ..$ ASV13: num [1:30] 1 0 7 0 1 0 1 0 0 0 ...
#>   ..$ ASV14: num [1:30] 0 0 0 2 1 0 0 0 2 0 ...
#>   ..$ ASV15: num [1:30] 1 0 11 0 0 1 0 0 0 1 ...
#>   ..$ ASV16: num [1:30] 0 3 0 1 0 0 0 0 0 0 ...
#>   ..$ ASV17: num [1:30] 2 3 4 0 0 1 0 0 0 1 ...
#>   ..$ ASV18: num [1:30] 0 0 0 1 0 0 0 1 0 0 ...
#>   ..$ ASV19: num [1:30] 2 0 0 3 1 0 1 0 0 1 ...
#>   ..$ ASV20: num [1:30] 0 0 10 0 0 1 0 0 1 1 ...
#>   ..$ ASV21: num [1:30] 2 0 0 0 0 0 0 0 0 0 ...
#>   ..$ ASV22: num [1:30] 3 0 5 3 0 1 0 0 0 1 ...
#>   ..$ ASV23: num [1:30] 0 0 5 3 0 3 0 0 0 2 ...
#>   ..$ ASV24: num [1:30] 0 2 0 0 0 1 1 0 0 0 ...
#>   ..$ ASV25: num [1:30] 3 3 8 0 1 1 0 1 0 0 ...
#>   ..$ ASV26: num [1:30] 2 3 3 1 2 0 0 0 0 0 ...
#>   ..$ ASV27: num [1:30] 1 0 8 3 2 1 0 0 0 0 ...
#>   ..$ ASV28: num [1:30] 1 2 0 2 0 0 0 0 0 1 ...
#>   ..$ ASV29: num [1:30] 0 4 7 2 0 0 0 0 0 1 ...
#>   ..$ ASV30: num [1:30] 1 1 0 0 3 0 1 0 0 1 ...
#>   ..$ ASV31: num [1:30] 0 4 4 3 0 0 0 0 1 0 ...
#>   ..$ ASV32: num [1:30] 0 0 0 2 0 2 0 0 0 1 ...
#>   ..$ ASV33: num [1:30] 0 2 0 0 1 0 1 0 1 0 ...
#>   ..$ ASV34: num [1:30] 2 2 0 0 0 3 2 0 0 0 ...
#>   ..$ ASV35: num [1:30] 0 4 2 0 0 1 0 0 0 0 ...
#>   ..$ ASV36: num [1:30] 3 0 9 1 1 1 0 0 0 3 ...
#>   ..$ ASV37: num [1:30] 3 5 4 1 0 2 0 1 1 2 ...
#>   ..$ ASV38: num [1:30] 3 2 10 2 0 0 0 0 0 0 ...
#>   ..$ ASV39: num [1:30] 0 5 10 3 1 0 1 0 0 0 ...
#>   ..$ ASV40: num [1:30] 0 0 4 3 1 0 3 0 0 0 ...
#>  $ graph       :List of 10
#>   ..$ : num 40
#>   ..$ : logi FALSE
#>   ..$ : num [1:200] 1 2 3 4 5 35 36 37 38 39 ...
#>   ..$ : num [1:200] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ : NULL
#>   ..$ : NULL
#>   ..$ : NULL
#>   ..$ : NULL
#>   ..$ :List of 4
#>   .. ..$ : num [1:3] 1 0 1
#>   .. ..$ :List of 7
#>   .. .. ..$ name    : chr "Watts-Strogatz random graph"
#>   .. .. ..$ dim     : num 1
#>   .. .. ..$ size    : num 40
#>   .. .. ..$ nei     : num 5
#>   .. .. ..$ p       : num 0.1
#>   .. .. ..$ loops   : logi FALSE
#>   .. .. ..$ multiple: logi FALSE
#>   .. ..$ : Named list()
#>   .. ..$ : Named list()
#>   ..$ :<environment: 0x10e049f90> 
#>   ..- attr(*, "class")= chr "igraph"
#>  $ id          :'data.frame':	30 obs. of  3 variables:
#>   ..$ plot : Factor w/ 3 levels "1","2","3": 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ depth: Factor w/ 2 levels "1","2": 1 1 1 1 1 2 2 2 2 2 ...
#>   ..$ time : int [1:30] 1 2 3 4 5 6 7 8 9 10 ...

res <- rri_pipeline_st(
  ROS_flux = redoxrri_example$ROS_flux,
  Eh_stability = redoxrri_example$Eh_stability,
  micro_data = redoxrri_example$micro_data,
  graph = redoxrri_example$graph,
  id = redoxrri_example$id,
  group_cols = c("plot", "depth"),
  scale_by = c("plot", "depth"),
  direction_phys = "auto",
  direction_anchor_phys = "FvFm",
  direction_soil = "auto",
  direction_anchor_soil = "Eh"
)

head(res$row_scores)
#>      Physio      Soil     Micro       RRI Micro_abundance Micro_network
#> 1 0.1900620 0.3555494 0.9404406 0.3300084       0.9404406           0.5
#> 2 0.3453262 0.4610492 0.7977234 0.4426647       0.7977234           0.5
#> 3 1.0000000 1.0000000 0.0000000 0.8891401       0.0000000           0.5
#> 4 0.3827148 0.3853520 0.8895645 0.4629757       0.8895645           0.5
#> 5 0.0000000 0.0000000 1.0000000 0.0000000       1.0000000           0.5
#> 6 0.9897659 1.0000000 0.0000000 0.8818604       0.0000000           0.5
#>   Micro_mfa
#> 1        NA
#> 2        NA
#> 3        NA
#> 4        NA
#> 5        NA
#> 6        NA
head(res$row_scores_comp)
#>      Physio      Soil     Micro       RRI
#> 1 0.1278973 0.2392577 0.6328450 0.3300084
#> 2 0.2152774 0.2874194 0.4973032 0.4426647
#> 3 0.5000000 0.5000000 0.0000000 0.8891401
#> 4 0.2308805 0.2324715 0.5366480 0.4629757
#> 5 0.0000000 0.0000000 1.0000000 0.0000000
#> 6 0.4974283 0.5025717 0.0000000 0.8818604